Use double precision for time measurements

+++ programs/keyboardController/LinearTrajectoryThread.hpp
@@ -32,7 +32,7 @@ protected:
 private:
     ICartesianControl * iCartesianControl;
     ICartesianTrajectory * iCartesianTrajectory;
-    int startTime;
+    double startTime;
     mutable yarp::os::Mutex mutex;
 };
